+++ tests/Neo.L2.Abstractions.UnitTests/UT_NeoHubDeployReport.cs
@@ -195,6 +195,8 @@ public class UT_NeoHubDeployReport
                 + "SettlementClient / TransactionSender / HasForcedInclusionFinalizer / "
                 + "HasSettlementClient / HasTransactionSender / IsSettlementEnabled / "
                 + "L1FinalityDepth / HasL1RpcEndpoint / ExpectedNetwork / "
+                + "HasSettlementManagerHash / HasForcedInclusionHash / "
+                + "HasSharedBridgeHash / HasMessageRouterHash / "
                 + "DepositSourceReadyCount / DepositSourceReservedCount / "
                 + "DepositSourceSoftConsumedCount / IsMetricsEnabled / MetricsConfiguredPort / "
                 + "MetricsBindAddress / ForcedInclusionDeploymentHeight / "